stock_predictor/requirements.txt
yfinance
pandas
numpy
python-dotenv
requests
newsapi-python
matplotlib